Add function to generate correct spline and main function
Change the way the original spline is calculated from the base_spline
and is returned as the solution of `Mx = y` in an array.

Create main() client function that receives relevant parameters and
plots the final graph with the points distorted by noise and the spline
calculated from those points by our function.

Signed-off-by: Victor Hugo M. Pinto <victor2142@gmail.com>